FE stats ugly, but works

+++ internal/models/statistics.go
@@ -5,7 +5,7 @@ type Statistics struct {
 	TotalSubmissions  int      `db:"-" json:"total_submissions"`
 	UniquePlayerCount int      `db:"-" json:"unique_player_count"`
 	MaxTier           *int     `db:"-" json:"max_tier"`
-	AverageTier       *float64 `db:"-" json:"average_tier,omitempty"`
+	AverageTier       *int     `db:"-" json:"average_tier,omitempty"`
 	PercentageTotal   *float64 `db:"-" json:"percentage_total,omitempty"`
 	PercentageUnique  *float64 `db:"-" json:"percentage_unique,omitempty"`
 }
